requirements.txt
Django==3.2.25
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.0
Pillow==10.0.1
djongo==1.3.6
pymongo==4.6.0
mongoengine==0.27.0
setuptools<81
